Refactor: rename `decoder` to `decoders` and `tokenizer` to `encoders`, update modules, structs, tests, and comments for consistency.
