+++ crates/bpetok/src/decoders/mod.rs
@@ -1,12 +1,12 @@
 //! # Token Decoders
 
 pub mod byte_decoder;
-pub mod context;
+pub mod decode_context;
 pub mod dictionary_decoder;
 pub mod pair_decoder;
 pub mod parallel_decoder;
 pub mod token_decoder;
 
-pub use context::*;
+pub use decode_context::*;
 pub use dictionary_decoder::*;
 pub use token_decoder::*;
