+++ crates/bpetok/src/training/trainer.rs
@@ -329,9 +329,9 @@ impl<T: TokenType, C: CountType> Ord for MergeJob<T, C> {
 
 #[cfg(test)]
 mod tests {
-    use crate::decoder::token_decoder::TokenDecoder;
-    use crate::tokenizer::TokenEncoder;
-    use crate::tokenizer::unified_encoder::UnifiedVocabEncoder;
+    use crate::decoders::token_decoder::TokenDecoder;
+    use crate::encoders::TokenEncoder;
+    use crate::encoders::unified_encoder::UnifiedVocabEncoder;
     use crate::training::trainer::{BPETokenVocabTrainer, MergeJob, TrainResults};
     use crate::types::{check_is_send, check_is_sync};
     use crate::vocab::unified_vocab::UnifiedTokenVocab;
